refactor(command-visualizer): split css-cascade.ts under 250-line limit
Extract the flash-fade section (FADE_FRAC, DELETE/ADD/REFERENCE_FLASH_STYLES,
flashFadeKeyframes, flashFadeRules) into render/css-cascade-flash.ts. Shared
pct() formatter is exported from css-cascade.ts and reused (no duplication).
Pure extraction; rendered CSS byte-identical.
